pyproject.toml
[project]
name = "ai-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.136.0",
    "google-genai>=1.73.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "tavily-python>=0.7.23",
    "uvicorn>=0.44.0",
]
